Introduce timer to disconnect peers stalling block download
This commit fixes a long standing bug which previously, would cause a
syncing node to get "stuck", halting all forward progress with respect
to keeping up with the main chain.

There are many nodes on the network which claim to be a fully
functioning full-node, but don't actually properly implement the entire
p2p protocol. For this particular case, after receiving a "getdata"
requesting blocks, these nodes fail to respond with a "block" message. This would
cause btcd to stall block download indefinitely, waiting for a response
which never arrives. Since btcd currently syncs new blocks via a single
syncPeer, it would never attempt to seek to new, active peers to
download the next block from. To fix this issue, we now disconnect peers
that we deem are stalling new block download after a specified timeout.

This fixes #486 and #229.