Implement optimized utxo set and view.
This removes the dependency on the transaction index for looking up
input transaction scripts and instead implements a utxo set that, as it
name implies, provides access specifically to the unspent transaction
outputs.  The serialized utxo entries use a domain specific compression
algorithm that understands the standard transasction scripts in order to
significantly compress them.

To go along with this, the chain code has been reworked to make the view
a rolling view as opposed to solely being a static storage mechanism.
This change is basically required for efficiently updating the utxo set
without having to calculate spent and added utxos twice since block
connection is checked before it is actually updated.  This also paves
the way for a utxo cache which will provide even more signficant
performance improvements.

In addition to providing significantly faster access to the inputs
needed to validate transactions, this has several other benefits such as
paving the way for block pruning and various proofs.

Finally, full tests for all compression and serialization code have been
added and the old utxostore data set used in the tests has been updated
to use the new utxo serialization format.  The reduced the data from
1.4M to 36K.