blockchain: Correct stxo resurrection code.
This corrects the resurrection code for spent txouts and adds a test to
trigger the original failing case.

The issue was caught by an assertion and was that the current state of
the utxo view needs to be checked before the in-flight map since
otherwise multiple inputs from the same transaction when the transaction
is not fully spent use the incorrect tx version.