blockchain: Correct unordered utxo serialization.
This corrects the utxo entry serialization to account for out of order
packed indexes in the outputs and adds a test to trigger the original
failing case.

In the normal case, the packed outputs are always in order, however when
resurrecting transactions they are packed into the outputs backwards, so
the serialization code needs to ensure they are properly ordered in
ascending order prior to serializing them.

This solution here isn't particuarly elegant, but the internal structure
data struct used to represent the entries need to be changed in order to
handle this better, so for now, just sort them before serializing.