blockchain: Add best hash to UtxoViewpoint.
This adds tracking for the current best hash a UtxoViewpoint represents
along with an assert in checkConnectBlock to enforce its correctness.

In addition to helping ensure correctness of the view's state, it is
useful for the upcoming cache as well since it will have its own best
hash as well.