txscript: optimize sigcache lookup
Profiles discovered that lookups into the signature cache included an
expensive comparison to the stored `sigInfo` struct. This lookup had
the potential to be more expensive than directly verifying the
signature itself!

To remedy this poor performance, the SigCache’s look up key is now
instead a fixed sized 32-byte hash, which is the result of
SHA-256(nonce || sigHash || sig || pubKey). This nonce value is
randomly generated upon initialization.

Additionally, the default value of the maximum number of entries
within the SigCache as been doubled due to the reduction in the
size of a cache entry.

Fixes #575.