server: change the scope of enablebanning
This update disables all ban score related logic unless banning is 
enabled. Previously, with enablebanning set to false, the ban score was
still calculated and warning messages were emitted if the score reached
the warning level, it was just the actual banning that was bypassed.
With this change, calculation and logging of ban score is fully disabled
unless banning is explicitly enabled.