database: Return empty slice on existing keys.
This modifies the internal treaps to make it such that setting a key to
a nil value creates an empty byte slice instead.

In turn, this ensures that an empty byte slice is returned for keys that
exist as opposed to nil.  This allows the caller to easily distinguish
between a key that doesn't exist and one that does from looking at the
returned value.

Also, this mirrors the behavior of the underlying leveldb interface.
Previously it was possible to get nil back for existing keys when they
were in the cache while they would return empty byte slices after being
flushed to disk.

Tests have been added to both the internal treaps and the database
interface itself to ensure the expeted behavior.