Replace a lot of big.Int math with nfield.
This changeset makes a bunch of operations on the secp256k1 curve faster by using a custom uint32-based math instead of using big.Int for calculating exponents. This is very similar to field.go, but is on the prime N, not P.

Preliminary indications are that there's about a 20% speed increase on signature verification.