rpcserver: Improve getwork handling.
This modifies a few things to improve the getwork RPC call handling.

First, it corrects the calculation of the getworkDataLen constant to
work with any size block header.  The current code calculated the
correct value for the current block header size, so the value of the
constant is not being changed, however the code was not accurate for an
arbitrary header size.

This is the case because the padding for the blake256 hash algorithm
requires a single 1 bit after the payload to be hashed in addition to
the 64 bits which encode the length of the message which is 65 bits.  It
also requires the final bit of the padding prior to the encoded message
length to be a 1 and to fall on bit 448.  This ensures the final block
is 64 bytes.

For a concrete example, the current code would incorrectly calculate
that a 184 block header only needs a getwork data length of 192, instead
of the 256 that is really required.

Second, because the block header size is constant and can't change while
the daemon is running, calculate the appropriate padding for the block
header size at init time rather than hard coding it directly in the get
work function.  This also means new memory doesn't need to be allocated
for the constant bytes on every invocation as it was previously doing.

Finally, preallocate the full size of the getwork buffer instead of
using append which incurs the overhead of additional memory copies.
This is how it's done in upstream btcsuite and I'm not sure why it was
changed to be less efficient in Decred.