Fix coin supply counter to reduce work and tax subsidy based on voters
The first pass for the getCoinSupply command did not accurately reduce
the work and tax subsidies based on the number of voters. So the
coinsupply was slightly high due to blocks that had less than 5 voters
not being properly reduced to match the proper total block reward. Looks
like this reduces current coin supply count by roughly 1507 DCR