Add lookup of ticket commitments to addrindex
Ticket commitments to a future address were ignored under the
old code. Now they are deserialized from their pkScripts and
stored in the address index of the database. This corrects some
wallet syncing behaviour where usage of addresses failed to
be determined with existsaddress. This has been tested for
correctness with several simnet blockchains.