peer: Remove extraneous hasTimestamp check.
The upstream project didn't have timestamps in address messages until
a certain protocol version, however that doesn't apply to Decred.  As a
result, it looks like the check for it was updated to just hard code it to
true.  This was accurate, but there is no point in setting something to
true and then turn around and check if it's true one line later.

Consequently, this just removes the check altogether.