server: Drain relay notification channel.
A new channel to handle inventory relaying asynchronous to the mempool
was added when the mempoolConfig struct was introduced in commit
83bcfea271da8b0c890ede7ef374c5a74ec5aad7.

However, the new channel was not also added to the cleanup code which
drains the channels and, as a result, it led to the possibility for the
server to hang during shutdown.

This commit resolves that issue.