chaincfg: Use hex strings for PoW limits.
Rather than having separate variables for each network's proof-of-work
limit parameter, introduce a new function to convert a hex string to a
big.Int and make use of it.  This has a couple of benefits such as being
easier to visually verify the limit at a glance and removing the need to
skip between two different areas of the file when looking at the
parameters.