wire: Don't allocate new readers in benchmarks.
This modifies the benchmarks in the wire package to avoid creating a new
reader for each iteration.  This is useful since it means that showing
the memory allocations will only show the function under test instead of
the allocation for the benchmark setup as well.

The following is a before and after comparison of the allocations
with the benchmarks that did not change removed:

benchmark            old allocs     new allocs     delta
----------------------------------------------------------
ReadVarInt1          2              1              -50.00%
ReadVarInt3          2              1              -50.00%
ReadVarInt5          2              1              -50.00%
ReadVarInt9          2              1              -50.00%
ReadVarStr4          4              3              -25.00%
ReadVarStr10         4              3              -25.00%
ReadOutPoint         2              1              -50.00%
ReadTxOut            4              3              -25.00%
ReadTxIn             6              5              -16.67%
DeserializeTx        33428          33427          -0.01%
ReadBlockHeader      8              7              -12.50%