wire: Use large tx for deserialize benchmark.
This modifies the transaction used the benchmark for deserializing
transactions to use what is often referred to as the megatransaction
since it is the largest Bitcoin transaction mined to date.  It consists
of 5569 inputs and 1 output.  Its hash is
bb41a757f405890fb0f5856228e23b715702d714d59bf2b1feb70d8b2b4e3e08.

This is being done so the benchmark is testing more of a worst-case
scenario which is a better candidate for identifying and testing
improvements.

The following benchmark results shows the how much more intensive this
transaction is over the mock transaction it is replacing:

old: DeserializeTx     1798 ns/op       376 B/op      16 allocs/op
new: DeserializeTx  5022304 ns/op   1672827 B/op   33429 allocs/op